+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Missing values in the version history modal for Suppressed from discovery and Staff suppressed fields. Fixed UIIN-3277.
 * Enable MARC-related options for shared MARC instance on member tenant. Fixes UIIN-3292.
 * ECS FOLIO instances: Inventory version history feature toggle/icon. Refs UIIN-3284.
+* Refactor `useAuditDataQuery` hooks to use `useInfiniteQuery` to implement loading indicator to version history panes. Refs UIIN-3286.
 
 ## [13.0.0](https://github.com/folio-org/ui-inventory/tree/v13.0.0) (2025-03-14)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.12...v13.0.0)